Moved specific parsers to their corresponding class - Replaced relative imports with absolute imports
