+++ tibiawikisql/models/__init__.py
@@ -1,9 +1,9 @@
-from .achievement import Achievement
-from .charm import Charm
-from .creature import Creature, CreatureDrop
-from .house import House
-from .imbuement import Imbuement, ImbuementMaterial
-from .item import Item, ItemAttribute, Key
-from .npc import Npc, NpcDestination, NpcSpell, NpcBuyOffer, NpcOffer, NpcSellOffer, RashidPosition
-from .quest import Quest, QuestDanger, QuestReward
-from .spell import Spell
+from tibiawikisql.models.achievement import Achievement
+from tibiawikisql.models.charm import Charm
+from tibiawikisql.models.creature import Creature, CreatureDrop
+from tibiawikisql.models.house import House
+from tibiawikisql.models.imbuement import Imbuement, ImbuementMaterial
+from tibiawikisql.models.item import Item, ItemAttribute, Key
+from tibiawikisql.models.npc import Npc, NpcDestination, NpcSpell, NpcBuyOffer, NpcOffer, NpcSellOffer, RashidPosition
+from tibiawikisql.models.quest import Quest, QuestDanger, QuestReward
+from tibiawikisql.models.spell import Spell
