+++ tibiawikisql/models/imbuement.py
@@ -1,15 +1,89 @@
 import re
 
 from tibiawikisql import schema, abc
-from tibiawikisql.utils import parse_effect
 
 astral_pattern = re.compile(r"\s*([^:]+):\s*(\d+),*")
+effect_pattern = re.compile(r"Effect/([^|]+)\|([^}|]+)")
 
 
 def parse_astral_sources(content: str):
+    """
+    Parses the astral sources of an imbuement.
+
+    Parameters
+    ----------
+    content: A string containing astral sources.
+
+    Returns
+    -------
+    :class:`dict[str,int]`:
+        A dictionary containing the material name and te amount required.
+    """
     materials = astral_pattern.findall(content)
     if materials:
         return {item: int(amount) for (item, amount) in materials}
+
+
+def parse_effect(effect):
+    """Parses TibiaWiki's effect template into a string effect.
+
+    Parameters
+    ----------
+    effect: class:`str`
+        The string containing the template.
+
+    Returns
+    -------
+    :class:`str`:
+        The effect string.
+    """
+    m = effect_pattern.search(effect)
+    category, amount = m.groups()
+    if category == "Bash":
+        return f"Club fighting +{amount}"
+    if category == "Chop":
+        return f"Axe fighting +{amount}"
+    if category == "Slash":
+        return f"Sword fighting +{amount}"
+    if category == "Precision":
+        return f"Distance fighting +{amount}"
+    if category == "Blockade":
+        return f"Shielding +{amount}"
+    if category == "Epiphany":
+        return f"Magic level +{amount}"
+    if category == "Scorch":
+        return f"Fire damage {amount}"
+    if category == "Venom":
+        return f"Earth damage {amount}"
+    if category == "Frost":
+        return f"Ice damage {amount}"
+    if category == "Electrify":
+        return f"Energy damage {amount}"
+    if category == "Reap":
+        return f"Death damage {amount}"
+    if category == "Vampirism":
+        return f"Life leech {amount}"
+    if category == "Void":
+        return f"Mana leech {amount}"
+    if category == "Strike":
+        return f"Critical {amount}"
+    if category == "Lich Shroud":
+        return f"Death protection {amount}"
+    if category == "Snake Skin":
+        return f"Earth protection {amount}"
+    if category == "Quara Scale":
+        return f"Ice protection {amount}"
+    if category == "Dragon Hide":
+        return f"Fire protection {amount}"
+    if category == "Cloud Fabric":
+        return f"Energy protection {amount}"
+    if category == "Demon Presence":
+        return f"Holy protection {amount}"
+    if category == "Swiftness":
+        return f"Speed +{amount}"
+    if category == "Featherweight":
+        return f"Capacity +{amount}"
+    return f"{category} {amount}"
 
 
 class Imbuement(abc.Row, abc.Parseable, table=schema.Imbuement):
